requirements.txt
# Core ML and Data Science
tensorflow>=2.13.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Trading and Financial Analysis
alpaca-py>=0.8.2
ta-lib>=0.4.28
yfinance>=0.2.28

# Deep Learning Enhancements
keras>=2.13.1

# Data Processing and Utilities
PyYAML>=6.0.1
python-dateutil>=2.8.2
pytz>=2023.3

# Numerical and Scientific Computing
scipy>=1.11.2

# Plotting and Visualization (optional but recommended)
matplotlib>=3.7.2
seaborn>=0.12.2

# Logging and Monitoring
logging>=0.4.9.6

# Development Tools
jupyter>=1.0.0
ipython>=8.14.0